requirements.txt

THIS CRASH GDAL!!!
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
descartes==1.1.0
dicttoxml==1.7.4
entrypoints==0.3
Fiona==1.8.4
GDAL==2.3.3
geopandas==0.4.1
gpxpy==1.3.5
idna==2.8
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
mapclassify==2.0.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
olefile==0.46
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pathlib==1.0.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.7.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyproj==1.9.6
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.3
requests==2.22.0
Rtree==0.8.3
scipy==1.3.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplekml==1.3.0
six==1.12.0
SQLAlchemy==1.3.7
SRTM.py==0.3.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
whitebox==0.9.0
widgetsnbextension==3.5.1